refactor: Redis 캐시 설정 및 연결 구성 수정 (#434)
* feat: Redis 캐시 설정 및 연결 구성 추가

* refactor: 레디스 설정 파일 분리 파일 삭제
